requirements.txt
autopep8==1.5.7
click==8.0.1
cycler==0.10.0
flake8==3.9.2
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
numpy==1.21.2
Pillow==8.3.2
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
